// std-toolkit/src/db/dynamodb/client/generated/types.ts
interface _AttributeValue {
  /**
   * An attribute of type String. For example:
   * "S": "Hello"
   */
  S?: string;
  /**
   * An attribute of type Number. For example:
   * "N": "123.45"
   * Numbers are sent across the network to DynamoDB as strings, to maximize compatibility
   * across languages and libraries. However, DynamoDB treats them as number type attributes
   * for mathematical operations.
   */
  N?: string;
  /**
   * An attribute of type Binary. For example:
   * "B": "dGhpcyB0ZXh0IGlzIGJhc2U2NC1lbmNvZGVk"
   */
  B?: Uint8Array | string;
  /**
   * An attribute of type String Set. For example:
   * "SS": ["Giraffe", "Hippo" ,"Zebra"]
   */
  SS?: Array<string>;
  /**
   * An attribute of type Number Set. For example:
   * "NS": ["42.2", "-19", "7.5", "3.14"]
   * Numbers are sent across the network to DynamoDB as strings, to maximize compatibility
   * across languages and libraries. However, DynamoDB treats them as number type attributes
   * for mathematical operations.
   */
  NS?: Array<string>;
  /**
   * An attribute of type Binary Set. For example:
   * "BS": ["U3Vubnk=", "UmFpbnk=", "U25vd3k="]
   */
  BS?: Array<Uint8Array | string>;
  /**
   * An attribute of type Map. For example:
   * "M": {"Name": {"S": "Joe"}, "Age": {"N": "35"}}
   */
  M?: Record<string, AttributeValue>;
  /**
   * An attribute of type List. For example:
   * "L": [ {"S": "Cookies"} , {"S": "Coffee"}, {"N": "3.14159"}]
   */
  L?: Array<AttributeValue>;
  /**
   * An attribute of type Null. For example:
   * "NULL": true
   */
  NULL?: boolean;
  /**
   * An attribute of type Boolean. For example:
   * "BOOL": true
   */
  BOOL?: boolean;
}

export type AttributeValue =
  | (_AttributeValue & { S: string })
  | (_AttributeValue & { N: string })
  | (_AttributeValue & { B: Uint8Array | string })
  | (_AttributeValue & { SS: Array<string> })
  | (_AttributeValue & { NS: Array<string> })
  | (_AttributeValue & { BS: Array<Uint8Array | string> })
  | (_AttributeValue & { M: Record<string, AttributeValue> })
  | (_AttributeValue & { L: Array<AttributeValue> })
  | (_AttributeValue & { NULL: boolean })
  | (_AttributeValue & { BOOL: boolean });
// ...
export interface BatchWriteItemInput {
  /**
   * A map of one or more table names or table ARNs and, for each table, a list of
   * operations to be performed (DeleteRequest or PutRequest). Each
   * element in the map consists of the following:
   * DeleteRequest - Perform a DeleteItem operation on the
   * specified item. The item to be deleted is identified by a Key
   * subelement:
   * Key - A map of primary key attribute values that uniquely
   * identify the item. Each entry in this map consists of an attribute name
   * and an attribute value. For each primary key, you must provide
   * all of the key attributes. For example, with a
   * simple primary key, you only need to provide a value for the partition
   * key. For a composite primary key, you must provide values for
   * both the partition key and the sort key.
   * PutRequest - Perform a PutItem operation on the
   * specified item. The item to be put is identified by an Item
   * subelement:
   * Item - A map of attributes and their values. Each entry in
   * this map consists of an attribute name and an attribute value. Attribute
   * values must not be null; string and binary type attributes must have
   * lengths greater than zero; and set type attributes must not be empty.
   * Requests that contain empty values are rejected with a
   * ValidationException exception.
   * If you specify any attributes that are part of an index key, then the
   * data types for those attributes must match those of the schema in the
   * table's attribute definition.
   */
  RequestItems: Record<string, Array<WriteRequest>>;
  ReturnConsumedCapacity?: ReturnConsumedCapacity;
  /**
   * Determines whether item collection metrics are returned. If set to SIZE,
   * the response includes statistics about item collections, if any, that were modified
   * during the operation are returned in the response. If set to NONE (the
   * default), no statistics are returned.
   */
  ReturnItemCollectionMetrics?: ReturnItemCollectionMetrics;
}
// ...
export type ConsistentRead = boolean;
// ...
export interface DeleteRequest {
  /**
   * A map of attribute name to attribute values, representing the primary key of the item
   * to delete. All of the table's primary key attributes must be specified, and their data
   * types must match those of the table's key schema.
   */
  Key: Record<string, AttributeValue>;
}
// ...
export interface GetItemInput {
  /**
   * The name of the table containing the requested item. You can also provide the
   * Amazon Resource Name (ARN) of the table in this parameter.
   */
  TableName: string;
  /**
   * A map of attribute names to AttributeValue objects, representing the
   * primary key of the item to retrieve.
   * For the primary key, you must provide all of the attributes. For example, with a
   * simple primary key, you only need to provide a value for the partition key. For a
   * composite primary key, you must provide values for both the partition key and the sort
   * key.
   */
  Key: Record<string, AttributeValue>;
  /**
   * This is a legacy parameter. Use ProjectionExpression instead. For more
   * information, see AttributesToGet in the Amazon DynamoDB Developer
   * Guide.
   */
  AttributesToGet?: Array<string>;
  /**
   * Determines the read consistency model: If set to true, then the operation
   * uses strongly consistent reads; otherwise, the operation uses eventually consistent
   * reads.
   */
  ConsistentRead?: boolean;
  ReturnConsumedCapacity?: ReturnConsumedCapacity;
  /**
   * A string that identifies one or more attributes to retrieve from the table. These
   * attributes can include scalars, sets, or elements of a JSON document. The attributes in
   * the expression must be separated by commas.
   * If no attribute names are specified, then all attributes are returned. If any of the
   * requested attributes are not found, they do not appear in the result.
   * For more information, see Specifying Item Attributes in the Amazon DynamoDB Developer
   * Guide.
   */
  ProjectionExpression?: string;
  /**
   * One or more substitution tokens for attribute names in an expression. The following
   * are some use cases for using ExpressionAttributeNames:
   * To access an attribute whose name conflicts with a DynamoDB reserved
   * word.
   * To create a placeholder for repeating occurrences of an attribute name in an
   * expression.
   * To prevent special characters in an attribute name from being misinterpreted
   * in an expression.
   * Use the # character in an expression to dereference
   * an attribute name. For example, consider the following attribute name:
   * Percentile
   * The name of this attribute conflicts with a reserved word, so it cannot be used
   * directly in an expression. (For the complete list of reserved words, see Reserved Words in the Amazon DynamoDB Developer
   * Guide). To work around this, you could specify the following for
   * ExpressionAttributeNames:
   * {"#P":"Percentile"}
   * You could then use this substitution in an expression, as in this example:
   * #P = :val
   * Tokens that begin with the : character are
   * expression attribute values, which are placeholders for the
   * actual value at runtime.
   * For more information on expression attribute names, see Specifying Item Attributes in the Amazon DynamoDB Developer
   * Guide.
   */
  ExpressionAttributeNames?: Record<string, string>;
}
// ...
export type Key = Record<string, AttributeValue>;
// ...
export type ProjectionExpression = string;
// ...
export interface PutRequest {
  /**
   * A map of attribute name to attribute values, representing the primary key of an item
   * to be processed by PutItem. All of the table's primary key attributes must
   * be specified, and their data types must match those of the table's key schema. If any
   * attributes are present in the item that are part of an index key schema for the table,
   * their types must match the index key schema.
   */
  Item: Record<string, AttributeValue>;
}
// ...
export type ReturnConsumedCapacity = 'INDEXES' | 'TOTAL' | 'NONE';
export type ReturnItemCollectionMetrics = 'SIZE' | 'NONE';
// ...
export type TableName = string;
// ...
export interface WriteRequest {
  /**
   * A request to perform a PutItem operation.
   */
  PutRequest?: PutRequest;
  /**
   * A request to perform a DeleteItem operation.
   */
  DeleteRequest?: DeleteRequest;
}